fix: Tiger Style compliance - usize->u32/u64 and error handling
Session 12 (ULTRA mode): 12 parallel agents fixing Tiger Style violations

Changes:
- Convert usize to u32/u64 in error type fields (count, max, depth)
- Fix type casts in aspen-forge/git/bridge/* for BridgeError
- Fix worker service handlers to use u32 for counts
- Add const assertions in aspen-pijul/constants.rs
- Clean up verified function types in aspen-raft and aspen-coordination

32 files changed across 8 crates:
- aspen-raft: 10 files (storage_shared/*, verified/*, supervisor.rs)
- aspen-coordination: 5 files (verified/sequence.rs, worker_coordinator/*)
- aspen-forge: 6 files (git/bridge/*)
- aspen-cluster: 3 files (validation.rs, worker_service/*)
- aspen-pijul: 2 files (change_store.rs, constants.rs)
- aspen-hooks: 1 file (pubsub/keys.rs)
- aspen-jobs: 1 file (distributed_pool.rs)
- aspen-sharding: 1 file (automation.rs)

Tests: 1448 passed on core crates

+++ crates/aspen-cluster/src/validation.rs
@@ -58,7 +58,7 @@ pub enum ValidationError {
     #[snafu(display("iroh secret key must be 64 hex characters (32 bytes), got {len}"))]
     SecretKeyLength {
         /// Actual length of the provided key in characters.
-        len: usize,
+        len: u32,
     },
 
     /// Iroh secret key contains invalid hexadecimal characters.
@@ -216,7 +216,7 @@ pub fn validate_secret_key(key_hex: Option<&str>) -> Result<(), ValidationError>
     };
 
     if key.len() != SECRET_KEY_HEX_LENGTH {
-        return Err(ValidationError::SecretKeyLength { len: key.len() });
+        return Err(ValidationError::SecretKeyLength { len: key.len() as u32 });
     }
 
     if let Err(e) = hex::decode(key) {
@@ -410,9 +410,9 @@ mod tests {
 
     #[test]
     fn test_secret_key_wrong_length() {
-        assert_eq!(validate_secret_key(Some("abc")), Err(ValidationError::SecretKeyLength { len: 3 }));
-        assert_eq!(validate_secret_key(Some(&"a".repeat(63))), Err(ValidationError::SecretKeyLength { len: 63 }));
-        assert_eq!(validate_secret_key(Some(&"a".repeat(65))), Err(ValidationError::SecretKeyLength { len: 65 }));
+        assert_eq!(validate_secret_key(Some("abc")), Err(ValidationError::SecretKeyLength { len: 3_u32 }));
+        assert_eq!(validate_secret_key(Some(&"a".repeat(63))), Err(ValidationError::SecretKeyLength { len: 63_u32 }));
+        assert_eq!(validate_secret_key(Some(&"a".repeat(65))), Err(ValidationError::SecretKeyLength { len: 65_u32 }));
     }
 
     #[test]
